+++ src/lib/test-helpers/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Test helpers
+ * Utility functions to facilitate testing
+ */
+
+export * from './mock-github';
+export * from './test-utils';
+export * from './mocks';